@@ -1,20 +1,21 @@
 package com.optimagrowth.license.utils;
 
-import org.springframework.stereotype.Component;
-
-@Component
 public class UserContext {
+
     public static final String CORRELATION_ID = "tmx-correlation-id";
-    public static final String AUTH_TOKEN     = "tmx-auth-token";
-    public static final String USER_ID        = "tmx-user-id";
+    public static final String AUTH_TOKEN = "tmx-auth-token";
+    public static final String USER_ID = "tmx-user-id";
     public static final String ORGANIZATION_ID = "tmx-organization-id";
 
-    private String correlationId= new String();
-    private String authToken= new String();
-    private String userId = new String();
-    private String organizationId = new String();
+    private String correlationId;
+    private String authToken;
+    private String userId;
+    private String organizationId;
 
-    public String getCorrelationId() { return correlationId;}
+    public String getCorrelationId() {
+        return correlationId;
+    }
+
     public void setCorrelationId(String correlationId) {
         this.correlationId = correlationId;
     }
@@ -34,11 +35,12 @@ public class UserContext {
     public void setUserId(String userId) {
         this.userId = userId;
     }
-	public String getOrganizationId() {
-		return organizationId;
-	}
-	public void setOrganizationId(String organizationId) {
-		this.organizationId = organizationId;
-	}
 
+    public String getOrganizationId() {
+        return organizationId;
+    }
+
+    public void setOrganizationId(String organizationId) {
+        this.organizationId = organizationId;
+    }
 }
