@@ -1,27 +1,26 @@
 package com.optimagrowth.license.utils;
 
-import org.springframework.util.Assert;
-
 public class UserContextHolder {
-    private static final ThreadLocal<UserContext> userContext = new ThreadLocal<UserContext>();
 
-    public static final UserContext getContext(){
+    private static final ThreadLocal<UserContext> userContext = new ThreadLocal<>();
+
+    public static UserContext getContext() {
         UserContext context = userContext.get();
-
         if (context == null) {
             context = createEmptyContext();
             userContext.set(context);
-
         }
-        return userContext.get();
+        return context;
     }
 
-    public static final void setContext(UserContext context) {
-        Assert.notNull(context, "Only non-null UserContext instances are permitted");
+    public static void setContext(UserContext context) {
+        if (context == null) {
+            throw new IllegalArgumentException("Only non-null UserContext instances are permitted");
+        }
         userContext.set(context);
     }
 
-    public static final UserContext createEmptyContext(){
+    public static UserContext createEmptyContext() {
         return new UserContext();
     }
 }
